API For PEXT, Deleted Everything Else For A Rewrite
There gota be a faster way to go from raw bitboards to a Move type
without that big of a performance hit. Also since I failed on Perft(3),
it is as good a time as any to redo this the better way

+++ crates/pext/src/attacks/bishops.rs
@@ -85,7 +85,7 @@ mod test_bishops {
         let blockers = "d4,d3,a2,h2,f4,g7,e6";
         for idx in idxs {
             println!("For Bishop on: {idx} & Blockers on {blockers}");
-            generate_bishop_attacks(idx.idx(), blockers.place()).print();
+            generate_bishop_attacks(idx.idx() as u64, blockers.place()).print();
             println!("---");
         }
     }
